[JENKINS-47105] Move the MemoryAssert test to its own JVM as in https://github.com/jenkinsci/workflow-cps-plugin/pull/178. Also deleting test workaround no longer needed in Jenkins 2.x.
