+++ src/test/java/org/jenkinsci/plugins/workflow/libs/LibraryAdderTest.java
@@ -24,7 +24,6 @@
 
 package org.jenkinsci.plugins.workflow.libs;
 
-import groovy.lang.MetaClass;
 import hudson.FilePath;
 import hudson.model.Job;
 import hudson.model.Result;
@@ -36,10 +35,6 @@ import hudson.plugins.git.extensions.GitSCMExtension;
 import hudson.slaves.WorkspaceList;
 import hudson.scm.SubversionSCM;
 import hudson.scm.ChangeLogSet;
-import java.lang.ref.WeakReference;
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Iterator;
@@ -49,8 +44,6 @@ import jenkins.plugins.git.GitSCMSource;
 import jenkins.plugins.git.GitSampleRepoRule;
 import jenkins.scm.impl.subversion.SubversionSCMSource;
 import jenkins.scm.impl.subversion.SubversionSampleRepoRule;
-import org.codehaus.groovy.reflection.ClassInfo;
-import org.codehaus.groovy.transform.ASTTransformationVisitor;
 import org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition;
 import org.jenkinsci.plugins.workflow.cps.GlobalVariable;
 import org.jenkinsci.plugins.workflow.cps.global.GrapeTest;
@@ -65,7 +58,6 @@ import org.junit.Rule;
 import org.jvnet.hudson.test.BuildWatcher;
 import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.JenkinsRule;
-import org.jvnet.hudson.test.MemoryAssert;
 import org.jvnet.hudson.test.TestExtension;
 
 public class LibraryAdderTest {
@@ -324,38 +316,6 @@ public class LibraryAdderTest {
         WorkflowRun b2 = (WorkflowRun) ra.run(ra.getOriginalScript(), Collections.singletonMap("trusted", originalScript.replace(originalMessage, "should not allowed"))).get();
         r.assertBuildStatusSuccess(b2); // currently do not throw an error, since the GUI does not offer it anyway
         r.assertLogContains(originalMessage, b2);
-    }
-
-    private static final List<WeakReference<ClassLoader>> LOADERS = new ArrayList<>();
-    public static void register(Object o) {
-        ClassLoader loader = o.getClass().getClassLoader();
-        System.err.println("registering " + o + " from " + loader);
-        LOADERS.add(new WeakReference<>(loader));
-    }
-    @Test public void loaderReleased() throws Exception {
-        sampleRepo.init();
-        sampleRepo.write("vars/leak.groovy", "def call() {" + LibraryAdderTest.class.getName() + ".register(this)}");
-        sampleRepo.git("add", "vars");
-        sampleRepo.git("commit", "--message=init");
-        GlobalLibraries.get().setLibraries(Collections.singletonList(new LibraryConfiguration("leak", new SCMSourceRetriever(new GitSCMSource(null, sampleRepo.toString(), "", "*", "", true)))));
-        WorkflowJob p = r.jenkins.createProject(WorkflowJob.class, "p");
-        p.setDefinition(new CpsFlowDefinition("@Library('leak@master') _; " + LibraryAdderTest.class.getName() + ".register(this); leak()", false));
-        r.assertBuildStatusSuccess(p.scheduleBuild2(0));
-        assertFalse(LOADERS.isEmpty());
-        try { // For Jenkins/Groovy 1. Cf. CpsFlowExecutionTest.loaderReleased.
-            Field f = ASTTransformationVisitor.class.getDeclaredField("compUnit");
-            f.setAccessible(true);
-            f.set(null, null);
-        } catch (NoSuchFieldException e) {}
-        { // ditto
-            MetaClass metaClass = ClassInfo.getClassInfo(LibraryAdderTest.class).getMetaClass();
-            Method clearInvocationCaches = metaClass.getClass().getDeclaredMethod("clearInvocationCaches");
-            clearInvocationCaches.setAccessible(true);
-            clearInvocationCaches.invoke(metaClass);
-        }
-        for (WeakReference<ClassLoader> loaderRef : LOADERS) {
-            MemoryAssert.assertGC(loaderRef, false);
-        }
     }
 
     @Issue({"JENKINS-38021", "JENKINS-31484"})
